missions/04-specialization/app/requirements.txt
chromadb==0.6.*
pyyaml
